refactor!(mocks): initialize both mocks together
It makes the code simpler, also the list of scripts, without much overhead.
