+++ e2e/helpers/edusharing-helper.ts
@@ -1,7 +1,6 @@
 import { Helper } from 'codeceptjs'
 import { Server } from 'http'
-
-import { EdusharingServer } from '../../src/edusharing-mock/server'
+import { EdusharingServer } from '../../src/mocks'
 
 class EdusharingHelper extends Helper {
   private edusharingMock: EdusharingServer
