+++ package.json
@@ -12,8 +12,7 @@
     "dev:frontend": "vite build --watch",
     "dev:databases": "docker compose up --detach --build",
     "dev:backend": "dotenvx run -- tsx --watch src/backend/index.ts",
-    "dev-edusharing": "dotenvx run -- tsx --watch src/mocks/edusharing/start-server.ts",
-    "dev-itslearning": "dotenvx run -- tsx --watch src/itslearning-mock/start-server.ts",
+    "dev-mocks": "dotenvx run -- tsx --watch src/mocks/start.ts",
     "format": "npm-run-all --sequential --continue-on-error \"format:*\"",
     "format:eslint": "eslint --fix",
     "format:prettier": "prettier . --write",
