+++ src/mocks/index.ts
@@ -1,1 +1,2 @@
 export * from './edusharing/server'
+export * from './itslearning'
